requirements.txt
discord.py
google-generativeai
python-dotenv
aiohttp
psutil
groq
PyNaCl
cryptography